refactor SidecarStats to telemetry namespace
